# gateway/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "US Government Data API",
    "description": "Unified REST API providing access to 45 US government data endpoints across 13 categories including cybersecurity vulnerability intelligence. Real-time data from NVD, CISA KEV, EPSS, NHTSA, FDA, CFPB, FEMA, NOAA, USGS, SEC, BLS, USDA, EPA, HUD, FAA, CBP, SAM.gov, CourtListener, CDC, WHO, Federal Register, Grants.gov, OpenCorporates, and more.",
    "version": "1.1.0",
    "contact": {
      "name": "MARTC03",
      "url": "https://apify.com/martc03"
    }
  },
  "servers": [
    {
      "url": "https://govdata-api.netlify.app",
      "description": "Production (Netlify Proxy)"
    }
  ],
  "paths": {
...
    "/api/v1/cyber/cve/{cveId}": {
      "get": {
        "operationId": "lookupCve",
        "summary": "Enriched CVE Lookup",
        "description": "Get comprehensive vulnerability intelligence for a CVE. Returns NVD details (CVSS scores, description, references), CISA KEV exploitation status, FIRST.org EPSS exploitation probability, and mapped MITRE ATT&CK techniques — all in a single call.",
        "tags": ["Cybersecurity"],
        "parameters": [
          { "name": "cveId", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^CVE-\\d{4}-\\d{4,}$" }, "description": "CVE identifier (e.g., CVE-2021-44228)" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/cyber/search": {
      "get": {
        "operationId": "searchCves",
        "summary": "Search CVE Database",
        "description": "Search the NVD CVE database by keyword, severity, date range, and KEV status.",
        "tags": ["Cybersecurity"],
        "parameters": [
          { "name": "keyword", "in": "query", "schema": { "type": "string" }, "description": "Search keyword (e.g., 'remote code execution', 'log4j')" },
          { "name": "severity", "in": "query", "schema": { "type": "string", "enum": ["LOW", "MEDIUM", "HIGH", "CRITICAL"] }, "description": "CVSS severity level" },
          { "name": "pubStartDate", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Published after (YYYY-MM-DD)" },
          { "name": "pubEndDate", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "Published before (YYYY-MM-DD)" },
          { "name": "hasKev", "in": "query", "schema": { "type": "boolean" }, "description": "Only return CVEs in CISA KEV catalog" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 10 }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
...
    "/api/v1/cyber/trending": {
      "get": {
        "operationId": "getTrendingCves",
        "summary": "Trending Critical Vulnerabilities",
        "description": "Get recently published critical and high severity CVEs from the NVD.",
        "tags": ["Cybersecurity"],
        "parameters": [
          { "name": "days", "in": "query", "schema": { "type": "integer", "default": 3 }, "description": "Look back N days (default 3)" },
          { "name": "severity", "in": "query", "schema": { "type": "string", "enum": ["HIGH", "CRITICAL"] }, "description": "Minimum severity (default: both HIGH and CRITICAL)" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    },
    "/api/v1/cyber/vendor/{vendor}": {
      "get": {
        "operationId": "getCvesByVendor",
        "summary": "CVEs by Vendor/Product",
        "description": "Search CVEs for a specific vendor and optionally a product. Includes KEV exploitation status for each result.",
        "tags": ["Cybersecurity"],
        "parameters": [
          { "name": "vendor", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Vendor name (e.g., Microsoft, Apache, Google)" },
          { "name": "product", "in": "query", "schema": { "type": "string" }, "description": "Product name (e.g., Windows, Log4j, Chrome)" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 }, "description": "Results to return" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/SuccessResponse" }, "502": { "$ref": "#/components/responses/ErrorResponse" } }
      }
    }
  },
  "components": {
    "responses": {
      "SuccessResponse": {
        "description": "Successful response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": true },
                "data": { "type": "object" },
                "meta": {
                  "type": "object",
                  "properties": {
                    "tool": { "type": "string" },
                    "latencyMs": { "type": "integer" },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string" },
                "meta": {
                  "type": "object",
                  "properties": {
                    "tool": { "type": "string" },
                    "latencyMs": { "type": "integer" },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "RapidAPIKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Key"
      }
    }
  },
  "security": [
    { "RapidAPIKey": [] }
  ],
...
}